added heregexp; now representing parses without loss of information
For example, a loop now has its own node instead of generating the same
AST that `while true` would generate. Don't worry; there'll be an
optimisation phase that removes these gratuitous nodes and replaces them
with their more general equivalents. The original parse is handy for
tools that need to know how the input was specified, not just a node
with equivalent semantics. For example, a text editor's code beautifier
would like to know that a heregex was not just an application of the
RegExp constructor on some concatenation. The optimisation phase will
make that transformation.