Use `.skip` instead of commenting out failing tests
Then you get a 'pending' count when the tests are run. This way, failing
tests are not forgotten about.

If a test contains syntax errors, the body of the test is commented out,
in order to be able to run any tests in the file in question at all.
These tests still increase the 'pending' count.

Best of all, about commented out tests actually passed!

Note:

- Some tests are still _partly_ commented out.
- I didn’t bother to convert splices.coffee.disabled.