Fix offset calculation for parsed nodes.
Closes #335.

This builds on #117 by ensuring that not only are line & column
information corrected to account for control characters inserted by the
preprocessor, but offset information is too. Unfortunately this expands
the hack that overrides the code generated by PEGjs, but it seemed the
cleanest and most performant way to solve the problem.