Fix duplicate vars at top level
To see the problem this patch fixes, look at the first line of
lib/cli.js:

```js
var $0, $0, ...
```

The Program node handler was attempting to deduplicate with `nub`, but
doing so over a list of JS.Identifiers, not a list of plain strings.

Since most callers of `declarationsNeededRecursive` are more interested
in the set of names than the identifier nodes, I changed it to do that.