Fix the reported line/column for syntax errors.
As reported at eventualbuddha/decaffeinate#54, CoffeeScriptRedux syntax
errors always point to line 1 column 1. This is very annoying when
trying to figure out what part of the file it actually can’t parse.
This updates the peg override to match the new signature for
`peg$computePosDetails` which now takes a `pos` argument.